requirements.txt
cachetools==5.3.3
fastapi==0.111.0
fastmcp==0.3.1
prometheus_client==0.20.0
slowapi==0.1.6
authlib==1.3.0               # For Azure AD OAuth
python-jose[cryptography]==3.3.0  # For JWT signing/verification
sqlalchemy[asyncio]==2.0.30  # For audit log ORM
aiosqlite==0.19.0            # For SQLite async DB
uvicorn[standard]==0.29.0    # For dev server & ASGI
